requirements-dev.txt
django
djangorestframework
django-filter

python-dotenv
psycopg2-binary